\name{geneswithsignalpeptide}
\alias{GenesWithSignalPeptide}
\title{Find genes that are predicted to encode a signal peptide containing protein.}
\description{Find genes that are predicted to encode a signal peptide.
          <br><br>
          The predictions are made with the <a href="http://www.cbs.dtu.dk/services/SignalP/">SignalP</a> program.}
\arguments{
\item{organism}{Organism
Provide one or more values. Use comma as a delimter.}
\item{min_sigp_sum_score}{SignalP-NN (neural network) provides four scores and a conclusion for each such that the conclusion  is 1 if the score meets a certain criteria and 0 otherwise.  The SignalP summary conclusion is the sum of the four conclusions generated by SignalP, thus four is the highest score possible.  We consider a summary conclusion score of three to be significant}
\item{min_sigp_d_score}{The D-score is the average of the S-mean and Y-max score from SignalP-NN}
\item{min_sigp_signal_probability}{Signal Probability from SignalP-HMM}
\item{any_or_all_sigp}{Union or intersection of advanced parameters.}
\item{o-fields}{Single valued attributes of the feature.
Provide one or more values. Use comma as a delimter.}
\item{o-tables}{Multi-valued attributes of the feature.
Provide one or more values. Use comma as a delimter.}
\item{.convert}{a logical value or a function that controls how the result of the method is returned. If this is a function, the character string or raw vector is passed to this function and it converts it appropriately. If this is a logical value and \code{TRUE}, then we attempt to convert the result based on its Content-Type returned by the Web server.  If this is \code{FALSE}, the value from the Web server is returned as is.}
\item{.url}{the URL for the Web request. This defaults to the correct value, but can be specified by the caller if the method is available at a different URL, e.g. locally or in a mirror server.}
\item{.json}{a logical value controlling whether to use the JSON or the XML version of the method}}
\value{text/xml
text/plain}
\author{}

