\name{WADLOptions}
\alias{AssembliesByLibrary.libraryIdAssemblies.options}
\alias{EstsByLocation.libraryId.options}
\alias{EstsByLibrary.libraryId.options}
\alias{EstsWithGeneOverlap.libraryIdGenes.options}
\alias{GenesByEcNumber.ec_number_pattern.options}
\alias{GenesByInterproDomain.domain_accession.options}
\alias{GenesByGoTerm.go_term.options}
\alias{GenesByESTOverlap.libraryIdGenes.options}
\alias{IsolateByTaxon.strain.options}
\alias{IsolateByProduct.product.options}
\alias{IsolateByStudy.study.options}
\title{Documentation for the Options for Parameters in WADL methods}
\description{This documents the variables used internally in
  the  R code to interface to a Web Service that was generated
  programmatically from a WADL
}
\author{Duncan Temple Lang}
\seealso{\code{\link[WADL]{makeFunctions}}}



